+++ weaver/wps_restapi/__init__.py
@@ -22,6 +22,7 @@ def includeme(config):
         config.add_route(**sd.service_api_route_info(sd.api_swagger_json_service, settings))
         config.add_route(**sd.service_api_route_info(sd.api_swagger_ui_service, settings))
         config.add_route(**sd.service_api_route_info(sd.api_versions_service, settings))
+        config.add_route(**sd.service_api_route_info(sd.api_conformance_service, settings))
         config.add_view(api.api_frontpage, route_name=sd.api_frontpage_service.name,
                         request_method="GET", renderer=OUTPUT_FORMAT_JSON)
         config.add_view(api.api_swagger_json, route_name=sd.api_swagger_json_service.name,
@@ -29,6 +30,8 @@ def includeme(config):
         config.add_view(api.api_swagger_ui, route_name=sd.api_swagger_ui_service.name,
                         request_method="GET", renderer="templates/swagger_ui.mako")
         config.add_view(api.api_versions, route_name=sd.api_versions_service.name,
+                        request_method="GET", renderer=OUTPUT_FORMAT_JSON)
+        config.add_view(api.api_conformance, route_name=sd.api_conformance_service.name,
                         request_method="GET", renderer=OUTPUT_FORMAT_JSON)
         config.add_notfound_view(api.not_found_or_method_not_allowed)
         config.add_forbidden_view(api.unauthorized_or_forbidden)